Refactored this assignment
Since a little app like this would be very useful to me, I changed
things around so that it would be fairly easy to add new pages.
I'm guessing that the assignment will be to add features to this.

In other words I refacored with extensibility and maintainability
much more in my mind.  Also I want things looking clean and neat
in this repository. I'll keep fiddling with it if I use it, and that
is definitely a good thing.